refactor(metrics): extract Prometheus impls into metric_collection/prometheus.rs
